+++ CHANGES.md
@@ -1,6 +1,7 @@
 # 0.3.0
 
 * Added POS tagging from CMU's ark-tweet-nlp for more accurate POS on intarwebspeek.
+* Added some basic tests using Expectations
 
 # 0.2.0